+++ OslSpreadsheet/Models/CellStyle.cs
@@ -9,6 +9,7 @@ namespace OslSpreadsheet.Models
         public string? BackgroundColor { get; set; }
         public string? FontName { get; set; }
         public double? FontSize { get; set; }
+        public bool WrapText { get; set; }
         public CellBorder? BorderTop { get; set; }
         public CellBorder? BorderBottom { get; set; }
         public CellBorder? BorderLeft { get; set; }